optimize socketsource/sink
